pyproject.toml
[project]
name = "class-assignmnet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.17",
    "chainlit>=1.0.0",
]
